Refactored to reduce amount of code.
The old code was very concerned about types and the goal was to try to
capture such errors at compile time.  However, the number of methods
became unweildly, so the code was refactored to use reflection to check
types.  This reduced the number of methods significantly and makes the
code resiliant to adverse conditions.